Find Git for Windows installations
Git for Windows's installer recommends not to put the git executables into the PATH environment variable. Therefore, 'git' will not be found when executed. The installer stores git's location in a registry key, however. This change will evaluate the registry on Windows systems and tries to find the full path to git.