Make Git.getOriginType more resilient
This change should suppress errors when there is not remote defined.

It also defends against the _remote_ case where someone might rename
origin to a different case (eg. Origin).

If there is no remote for the current branch, but there is at least one
remote set up for the repo, then use that to get the remote type.